+++ thesentinel/src/main/java/com/github/javezki/Sentinel.java
@@ -1,5 +1,8 @@
 package com.github.javezki;
 
+import com.github.javezki.Listeners.CommandListeners;
+import com.github.javezki.Listeners.CurrentEventListener;
+
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Activity;
