--- thesentinel/src/main/java/com/github/javezki/SentinelLog.java
@@ -1,8 +1,0 @@
-package com.github.javezki;
-
-public class SentinelLog {
-    
-    public SentinelLog() {
-        
-    }
-}
